requirements.txt
fastapi[all]==0.94.0
anyio==3.6.2
certifi==2022.12.7
chardet==4.0.0
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.94.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
orjson==3.8.10
psycopg2-binary==2.9.6
pydantic==1.10.6
pygame==2.3.0
pygame-gui==0.6.8
python-dotenv==0.21.0
python-i18n==0.3.9
python-multipart==0.0.6
PyYAML==6.0
requests==2.25.1
rfc3986==1.5.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.15
uvicorn==0.21.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.2
Werkzeug==2.2.3